requirements.txt
asyncpraw
dotenv
tiktoken
pandas
phidata
groq
uvicorn
fastapi